feat(sdk-lib-mpc): move types to types.ts
TICKET: WP-400
